+++ src/v3/data/projectDetails.js
@@ -1,0 +1,447 @@
+/** Ported from legacy Project.js (getContents + skillsBak). */
+export const PROJECT_DETAILS = {
+  "eco": {
+    "description": "Made in wordpress and woocommerce plugin. My job is to maintain the codes, enhance, debug the site. Make a custom plugin to work in woocommerce and additional features. Fix different bugs on design/layout. Fix website to make it more SEO friendly.",
+    "tags": [
+      "#wordpress",
+      "#woocommerce",
+      "#php",
+      "#html",
+      "#css",
+      "#js",
+      "#jquery",
+      "#pug",
+      "#nodejs",
+      "#seo",
+      "#debug",
+      "#enhancement",
+      "#sass",
+      "#jade",
+      "#plugin",
+      "#wordpressplugin",
+      "#admin",
+      "#mobileresponsive",
+      "#angularjs"
+    ],
+    "websiteUrl": "https://www.ecoshiftcorp.com/"
+  },
+  "sc": {
+    "description": "Made in .net system (kentico). Redesign the old layout from scratch from PSD Layout to HTML,CSS,JS. Make it friendly site to desktop and mobile.",
+    "tags": [
+      "#html",
+      "#css",
+      "#sass",
+      "#javascript",
+      "#jquery",
+      "#pug",
+      "#jade",
+      "#psdtohtml",
+      "#bootstrap",
+      "#mobileresponsive"
+    ],
+    "websiteUrl": "https://starcinema.abs-cbn.com/"
+  },
+  "kty": {
+    "description": "Made in .net system (kentico). Redesign the old layout from scratch from PSD Layout to HTML,CSS,JS. Make it friendly site to desktop and mobile.",
+    "tags": [
+      "#html",
+      "#css",
+      "#sass",
+      "#javascript",
+      "#jquery",
+      "#pug",
+      "#jade",
+      "#psdtohtml",
+      "#angular",
+      "#mobileresponsive"
+    ],
+    "websiteUrl": "https://thankyou.abs-cbn.com/"
+  },
+  "push": {
+    "description": "Made in .net system (kentico). Redesign the old layout from scratch from PSD Layout to HTML,CSS,JS. Make it friendly site to desktop and mobile.",
+    "tags": [
+      "#html",
+      "#psdtohtml",
+      "#pug",
+      "#jade",
+      "#nodejs",
+      "#debug",
+      "#javascript",
+      "#jquery",
+      "#socialmedia",
+      "#seo",
+      "#css",
+      "#sass",
+      "#mobileresponsive"
+    ],
+    "websiteUrl": "https://push.abs-cbn.com"
+  },
+  "corp": {
+    "description": "Enhance the website.",
+    "tags": [
+      "#html",
+      "#css",
+      "#sass",
+      "#javascript",
+      "#jquery",
+      "#pug",
+      "#jade",
+      "#psdtohtml",
+      "#mobileresponsive",
+      "#enhancement",
+      "#debug"
+    ],
+    "websiteUrl": "https://www.abs-cbn.com/corporate"
+  },
+  "careers": {
+    "description": "Enhance the careers website.",
+    "tags": [
+      "#html",
+      "#css",
+      "#sass",
+      "#javascript",
+      "#jquery",
+      "#pug",
+      "#jade",
+      "#psdtohtml",
+      "#enhancement",
+      "#debug"
+    ],
+    "websiteUrl": "https://careers.abs-cbn.com/"
+  },
+  "anc": {
+    "description": "ANC Form for 1 week event.",
+    "tags": [
+      "#html",
+      "#css",
+      "#sass",
+      "#javascript",
+      "#jquery",
+      "#pug",
+      "#jade",
+      "#psdtohtml",
+      "#enhancement",
+      "#debug"
+    ]
+  },
+  "kaiaq": {
+    "description": "KAIAQ Admin Template to be use.",
+    "tags": [
+      "#html",
+      "#css",
+      "#sass",
+      "#javascript",
+      "#jquery",
+      "#pug",
+      "#jade",
+      "#psdtohtml",
+      "#enhancement",
+      "#debug",
+      "#bootstrap"
+    ]
+  },
+  "asphaltnitro": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "https://www.gameloft.com/en/game/asphalt-nitro-mobile"
+  },
+  "asphalt8": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "https://www.gameloft.com/asphalt8/"
+  },
+  "cc": {
+    "description": "Enhance the site and fixed the problem from PHP, HTML, CSS, JS.",
+    "tags": [
+      "#codeigniter",
+      "#debug",
+      "#enhancement",
+      "#php"
+    ],
+    "websiteUrl": "https://support.gameloft.com/"
+  },
+  "dm": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "https://minionrush.com/"
+  },
+  "dh5": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "http://dungeonhunter5.com/"
+  },
+  "gl-forum": {
+    "description": "Enhance the site and fixed the problem from PHP, HTML, CSS, JS.",
+    "tags": [
+      "#codeigniter",
+      "#debug",
+      "#enhancement",
+      "#php"
+    ],
+    "websiteUrl": "https://forum.gameloft.com/"
+  },
+  "gor": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "https://www.gameloft.com/en/game/gods-of-rome"
+  },
+  "hoc": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "http://www.heroesoforderandchaosgame.com/index.php"
+  },
+  "sf": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "https://www.sniperfury.com/"
+  },
+  "tbc": {
+    "description": "Made it from scratch and from PSD Layout to HTML,CSS,JS. Friendly website from desktop to mobile and created a codeigniter system to make the repeatable codes in one for all the gamelofts 1 pager sites.",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#html",
+      "#php",
+      "#debug",
+      "#mobileresponsive",
+      "#css",
+      "#less",
+      "#nodejs"
+    ],
+    "websiteUrl": "https://www.gameloft.com/en-gb/game/blacklist"
+  },
+  "coll8or": {
+    "description": "All social media in one site by using the social medias API.",
+    "tags": [
+      "#socialmedia",
+      "#php",
+      "#codeigniter",
+      "#debug",
+      "#html",
+      "#api",
+      "#javascript",
+      "#jquery",
+      "#css",
+      "#less"
+    ],
+    "websiteUrl": "http://coll8or.com/"
+  },
+  "kinkcakes": {
+    "description": "About cakes.",
+    "tags": [
+      "#css",
+      "#less",
+      "#psdtohtml",
+      "#javascript",
+      "#jquery",
+      "#debug"
+    ],
+    "websiteUrl": "http://kinkcakes.com.ph/"
+  },
+  "zmr": {
+    "description": "News About Zenos websites and all good works of Zeno Company in every city and country.",
+    "tags": [
+      "#css",
+      "#sass",
+      "#psdtohtml",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#html",
+      "#wordpress",
+      "#wordpresstheme"
+    ],
+    "websiteUrl": "http://zenomediarelease.com/"
+  },
+  "zbceuit": {
+    "description": "About construction on italy-based company.",
+    "tags": [
+      "#css",
+      "#sass",
+      "#psdtohtml",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#html",
+      "#jade",
+      "#pug"
+    ],
+    "websiteUrl": "http://zbccostruzioni.it/"
+  },
+  "zgi": {
+    "description": "About investments and the members of zeno group investments.",
+    "tags": [
+      "#css",
+      "#sass",
+      "#psdtohtml",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#html",
+      "#jade",
+      "#pug",
+      "#laravel"
+    ],
+    "websiteUrl": "http://zenogroupinvestments.com/home"
+  },
+  "vigattintourism": {
+    "description": "Tourism site for the Philippines.",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#css",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#test",
+      "#enhancement",
+      "#seo",
+      "#mobileresponsive",
+      "#less"
+    ],
+    "websiteUrl": "https://www.vigattintourism.com/"
+  },
+  "vigattindeals": {
+    "description": "It is about a site with a coupon that can buy with very big discounts.",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#css",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#test",
+      "#paymentapi",
+      "#enhancement",
+      "#seo",
+      "#mobileresponsive",
+      "#less"
+    ],
+    "websiteUrl": "https://www.vigattindeals.com/"
+  },
+  "vigattintrade": {
+    "description": "It is a trading site like OLX. Buy and Sell Website.",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#css",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#test",
+      "#enhancement",
+      "#seo",
+      "#mobileresponsive",
+      "#less"
+    ],
+    "websiteUrl": "https://www.vigattintrade.com/"
+  },
+  "vigattin": {
+    "description": "The main site of all vigattin websites.",
+    "tags": [
+      "#codeigniter",
+      "#psdtohtml",
+      "#css",
+      "#javascript",
+      "#jquery",
+      "#debug",
+      "#test",
+      "#enhancement",
+      "#seo",
+      "#mobileresponsive",
+      "#less"
+    ],
+    "websiteUrl": "https://www.vigattin.com/"
+  },
+  "mb1": {
+    "description": "The MB AppKey feature front-end development spearheaded by me was all about laying down the groundwork for the initial module, including the design and full implementation of an entire base plan. Redux integration, unit tests, views, and subcomponents created a solid and reusable backbone. Once the core part was finished, it could be integrated easily by other developers across pages and modules—usually in less than five lines of code. Updates to the core propagate throughout the entire application.",
+    "tags": [
+      "#reactJS",
+      "#redux",
+      "#frontend",
+      "#unittesting",
+      "#API",
+      "#html",
+      "#css"
+    ],
+    "gallery": [
+      "/static/images/sites/mb2.png",
+      "/static/images/sites/mb3.png"
+    ]
+  }
+};
+
+export function getProjectDetails(projectId) {
+  if (!projectId) return null;
+  return PROJECT_DETAILS[projectId] || null;
+}